Updated PR addressing Fabian's comments
